satisfy https://github.com/filipdanic/json2csvexporter/issues/2#issuecomment-333463408
