+++ lib/CSVExportService.js
@@ -1,145 +1,25 @@
-'use strict';
+"use strict";
 
-function _classCallCheck(instance, Constructor) {
-  if (!(instance instanceof Constructor)) {
-    throw new TypeError("Cannot call a class as a function");
-  }
-}
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+exports.default = void 0;
 
-function _defineProperties(target, props) {
-  for (var i = 0; i < props.length; i++) {
-    var descriptor = props[i];
-    descriptor.enumerable = descriptor.enumerable || false;
-    descriptor.configurable = true;
-    if ("value" in descriptor) descriptor.writable = true;
-    Object.defineProperty(target, descriptor.key, descriptor);
-  }
-}
+var _WriterService = _interopRequireDefault(require("./WriterService"));
 
-function _createClass(Constructor, protoProps, staticProps) {
-  if (protoProps) _defineProperties(Constructor.prototype, protoProps);
-  if (staticProps) _defineProperties(Constructor, staticProps);
-  return Constructor;
-}
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
-/**
- * Provides an interface for mapping JS data types to a delimiter-separated value.
- * Creates a list of lists called rows and maps it to a Blob via toBlob().
- */
-var WriterService = /*#__PURE__*/function () {
-  /**
-   * Default constructors. Takes two optional params.
-   * @param  {string} delimiter - Delimiter character(s) to be used in the CSV.
-   * @param  {string} contentType - Type of file.
-   */
-  function WriterService() {
-    var delimiter = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : ',';
-    var contentType = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : 'text/csv';
+function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
 
-    _classCallCheck(this, WriterService);
+function _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }
 
-    this.delimiter = delimiter;
-    this.contentType = contentType;
-    this.rows = [[]];
-  }
-  /**
-   * Returns the current row
-   * @return {Array} - An array of values.
-   */
-
-
-  _createClass(WriterService, [{
-    key: "wrapWithQuotes",
-
-    /**
-     * Returns the input string
-     * @param  {string} string - The input string.
-     * @return {string} - A safe strings wrapped in quotes.
-     */
-    value: function wrapWithQuotes(string) {
-      var safeString = string.replace(/"/g, '""');
-      return "\"".concat(safeString, "\"");
-    }
-    /**
-     * @param {*} value
-     * @return {string}
-     */
-
-  }, {
-    key: "sanitizeValue",
-    value: function sanitizeValue(value) {
-      if (value === undefined || value === null || typeof value === 'function') {
-        return '';
-      }
-
-      return String(value);
-    }
-    /**
-     * Pushes a new value into the current row.
-     * @param  {*} value - The value to push.
-     */
-
-  }, {
-    key: "writeValue",
-    value: function writeValue(value) {
-      var stringValue = this.sanitizeValue(value);
-      var needsQuotes = stringValue.indexOf(this.delimiter) !== -1 || /"\r\n/.test(stringValue);
-      this.currentRow.push(needsQuotes ? this.wrapWithQuotes(stringValue) : stringValue);
-    }
-    /**
-     * Adds a en empty array to rows property. This maps to an empty line.
-     */
-
-  }, {
-    key: "writeLine",
-    value: function writeLine() {
-      this.rows.push([]);
-    }
-    /**
-     * Flatten rows to a String.
-     * @return {string} - A string representation of the rows..
-     */
-
-  }, {
-    key: "toString",
-    value: function toString() {
-      var _this = this;
-
-      return this.rows.map(function (row) {
-        return row.join(_this.delimiter);
-      }).reduce(function (content, row) {
-        return "".concat(content, "\r\n").concat(row);
-      });
-    }
-    /**
-     * Transform the rows into a Blob.
-     * @return {Object} - A representation of the rows in the form of
-     * a Blob. The returned Object is an instance of Blob, but typeof Object.
-     */
-
-  }, {
-    key: "toBlob",
-    value: function toBlob() {
-      return new Blob([this.toString()], {
-        type: this.contentType
-      });
-    }
-  }, {
-    key: "currentRow",
-    get: function get() {
-      return this.rows[this.rows.length - 1];
-    }
-  }]);
-
-  return WriterService;
-}();
+function _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); return Constructor; }
 
 /**
  * Provides an ExportService that uses WriterService to mentain an in-memory representation of the CSV file.
  * Uses downloadCSV to map the [data] via the Writer to a Blob and initiate download which triggers the download of the actual file in the browser.
  * Provides static methods for Singleton-like usage.
  */
-
 var CSVExportService = /*#__PURE__*/function () {
   /**
    * Default constructor. Takes an optional options param.
@@ -203,7 +83,7 @@ var CSVExportService = /*#__PURE__*/function () {
       var headerNames = optionsHeaders || {};
       var headers = optionsColumns || Object.getOwnPropertyNames(data[0]);
       var includeHeaders = optionsIncludeHeaders;
-      var writer = new WriterService(delimiter, contentType);
+      var writer = new _WriterService.default(delimiter, contentType);
 
       if (includeHeaders === undefined || includeHeaders) {
         headers.forEach(function (header) {
@@ -294,5 +174,4 @@ var CSVExportService = /*#__PURE__*/function () {
   return CSVExportService;
 }();
 
-module.exports = CSVExportService;
-//# sourceMappingURL=CSVExportService.js.map
+exports.default = CSVExportService;
